More papers May june 2024 Region 1 15 Marker Q 10
The one-dimensional (1D) array Clubs[] is used to store the names of 12 cricket clubs in a local
sports league.
The two-dimensional (2D) array Statistics[] is used to store, for each cricket club, the number
of:
• matches won
• matches drawn
• matches lost.
The 1D array Points[] is used to store the total number of points each cricket club has been
awarded.
The position of any cricket club’s data is the same in all three arrays. For example, the data in
index 2 of Statistics[] and index 2 of Points[] belongs to the cricket club in index 2 of
Clubs[]
The variable Matches stores the number of matches played by each team. Each team plays the
same number of matches.
Points are awarded for:
• a win – 12 points
• a draw – 5 points
• a loss – 0 points.
Write a program that meets the following requirements:
• allows the number of matches played to be input and stored, with a maximum of 22 matches
• validates the number of matches played
• allows the names of the cricket clubs to be input and stored
• allows the number of matches won, drawn and lost to be input and stored for each team
• validates the number of matches won, drawn or lost against the number of matches played
• asks the user to re-enter the number of matches won, drawn or lost if the total does not match
the number of matches played
• calculates and stores the total number of points for each club
• finds the cricket club or clubs with the highest number of points
• outputs the name or names of the winning club or clubs, the number of wins and the total
number of points awarded.
You must use pseudocode or program code and add comments to explain how your code works.
You do not need to declare any arrays or variables; you may assume that this has already been
done.
All inputs and outputs must contain suitable messages.